move date-parts.ts back into fake-date/
duplicate getOffsetMinutes to avoid bundle splitting
